Add card router and Event model
