+++ index.js
@@ -7,6 +7,7 @@ const app = express();
 const { Card } = require('./models/card');
 const mongoose = require('mongoose');
 const { MONGODB_URI, PORT, CLIENT_ORIGIN } = require('./config');
+const cardRouter = require('./routes/cards');
 
 app.use(
   morgan(process.env.NODE_ENV === 'production' ? 'common' : 'dev', {
@@ -22,84 +23,86 @@ app.use(
 
 app.use(express.json());
 
-app.get('/api', (req, res, next) => {
-  Card.find()
-    .then(response => {
-      res.json(response);
-    })
-    .catch(err => {
-      next(err);
-    });
-});
+app.use('/', cardRouter);
 
-app.post('/api/cards', (req, res, next) => {
-  if (!req.body.name) {
-    const err = new Error('Must include name');
-    err.status = 400;
-    return next(err);
-  }
+// app.get('/api', (req, res, next) => {
+//   Card.find()
+//     .then(response => {
+//       res.json(response);
+//     })
+//     .catch(err => {
+//       next(err);
+//     });
+// });
 
-  const newCard = {
-    name: req.body.name
-  };
+// app.post('/api/cards', (req, res, next) => {
+//   if (!req.body.name) {
+//     const err = new Error('Must include name');
+//     err.status = 400;
+//     return next(err);
+//   }
 
-  Card.create(newCard)
-    .then(response => {
-      if (response) {
-        res.json(response);
-      } else {
-        next();
-      }
-    })
-    .catch(next);
-});
+//   const newCard = {
+//     name: req.body.name
+//   };
 
-app.put('/api/cards/:id', (req, res, next) => {
-  if (req.body.id !== req.params.id) {
-    const err = new Error('Id must match in body and url params');
-    err.status = 400;
-    return next(err);
-  }
+//   Card.create(newCard)
+//     .then(response => {
+//       if (response) {
+//         res.json(response);
+//       } else {
+//         next();
+//       }
+//     })
+//     .catch(next);
+// });
 
-  const acceptedFields = ['name'];
-  const updateCard = {};
-  for (let field in acceptedFields) {
-    if (req.body[field]) {
-      updateCard[field] = req.body[field];
-    }
-  }
+// app.put('/api/cards/:id', (req, res, next) => {
+//   if (req.body.id !== req.params.id) {
+//     const err = new Error('Id must match in body and url params');
+//     err.status = 400;
+//     return next(err);
+//   }
 
-  if (req.body.xArray) {
-    Card.findById(req.params.id)
-      .then(response => {
-        updateCard.xArray = [...response.xArray, req.body.xArray];
-        console.log(updateCard);
-        return Card.findByIdAndUpdate(req.params.id, updateCard, { new: true });
-      })
-      .then(response => {
-        console.log(response);
-        if (response) {
-          res.json(response);
-        } else {
-          console.log('here');
-          next();
-        }
-      })
-      .catch(next);
-  } else {
-    Card.findByIdAndUpdate(req.params.id, updateCard, { new: true })
-      .then(response => {
-        if (response) {
-          res.json(response);
-        } else {
-          next();
-        }
-      })
-      .catch(next);
-  }
+//   const acceptedFields = ['name'];
+//   const updateCard = {};
+//   for (let field in acceptedFields) {
+//     if (req.body[field]) {
+//       updateCard[field] = req.body[field];
+//     }
+//   }
+
+//   if (req.body.xArray) {
+//     Card.findById(req.params.id)
+//       .then(response => {
+//         updateCard.xArray = [...response.xArray, req.body.xArray];
+//         console.log(updateCard);
+//         return Card.findByIdAndUpdate(req.params.id, updateCard, { new: true });
+//       })
+//       .then(response => {
+//         console.log(response);
+//         if (response) {
+//           res.json(response);
+//         } else {
+//           console.log('here');
+//           next();
+//         }
+//       })
+//       .catch(next);
+//   } else {
+//     Card.findByIdAndUpdate(req.params.id, updateCard, { new: true })
+//       .then(response => {
+//         if (response) {
+//           res.json(response);
+//         } else {
+//           next();
+//         }
+//       })
+//       .catch(next);
+//   }
 
   
-});
+// });
 
 app.use(function(req, res, next) {
   const err = new Error('Not Found');
